refactor(tests): consolidate OID4VC integration tests, remove ~1,469 lines of duplicates
