--- oid4vc/integration/tests/test_cross_wallet_compatibility.py
@@ -1,1383 +1,0 @@
-"""Cross-wallet compatibility tests for OID4VC.
-
-These tests discover interoperability bugs between Credo and Sphereon by:
-1. Issuing credentials to one client and verifying with another
-2. Testing format support differences
-3. Testing edge cases in algorithm negotiation
-4. Comparing selective disclosure behavior
-"""
-
-import asyncio
-import uuid
-
-import pytest
-
-from .test_config import MDOC_AVAILABLE  # noqa: F401
-
-
-def extract_credential(response, wallet_name: str) -> str:
-    """Safely extract credential from wallet response, skipping test if unavailable.
-
-    Args:
-        response: The HTTP response from wallet accept-offer call
-        wallet_name: Name of wallet for error messages (e.g., "Credo", "Sphereon")
-
-    Returns:
-        The credential string
-
-    Raises:
-        pytest.skip: If credential could not be obtained (infrastructure issue)
-    """
-    if response.status_code != 200:
-        pytest.skip(
-            f"{wallet_name} failed to accept offer (status {response.status_code}): {response.text}"
-        )
-
-    resp_json = response.json()
-    if "credential" not in resp_json:
-        pytest.skip(f"{wallet_name} did not return credential: {resp_json}")
-
-    return resp_json["credential"]
-
-
-# =============================================================================
-# Cross-Wallet Issuance and Verification Tests
-# =============================================================================
-
-
-@pytest.mark.asyncio
-async def test_issue_to_credo_verify_with_sphereon_jwt_vc(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    credo_client,
-    sphereon_client,  # noqa: ARG001
-):
-    """Issue JWT VC to Credo, then verify presentation from Credo via Sphereon-style request.
-
-    This tests whether credentials issued to Credo can be presented to a verifier
-    that uses Sphereon-compatible verification patterns.
-    """
-    # Step 1: Issue JWT VC credential to Credo
-    random_suffix = str(uuid.uuid4())[:8]
-    credential_supported = {
-        "id": f"CrossWalletCredential_{random_suffix}",
-        "format": "vc+sd-jwt",
-        "scope": "CrossWalletTest",
-        "proof_types_supported": {
-            "jwt": {"proof_signing_alg_values_supported": ["EdDSA", "ES256"]}
-        },
-        "format_data": {
-            "cryptographic_binding_methods_supported": ["did:key"],
-            "cryptographic_suites_supported": ["EdDSA"],
-            "vct": "CrossWalletCredential",
-            "claims": {
-                "name": {"mandatory": True},
-                "email": {"mandatory": False},
-            },
-        },
-        "vc_additional_data": {"sd_list": ["/name", "/email"]},
-    }
-
-    credential_config_response = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create", json=credential_supported
-    )
-    config_id = credential_config_response["supported_cred_id"]
-
-    did_response = await acapy_issuer_admin.post(
-        "/wallet/did/create", json={"method": "key", "options": {"key_type": "ed25519"}}
-    )
-    issuer_did = did_response["result"]["did"]
-
-    exchange_request = {
-        "supported_cred_id": config_id,
-        "credential_subject": {
-            "name": "Cross Wallet Test",
-            "email": "cross@wallet.test",
-        },
-        "did": issuer_did,
-    }
-
-    exchange_response = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create", json=exchange_request
-    )
-    exchange_id = exchange_response["exchange_id"]
-
-    offer_response = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange_id}
-    )
-    credential_offer_uri = offer_response["credential_offer"]
-
-    # Credo accepts the offer
-    accept_offer_request = {
-        "credential_offer": credential_offer_uri,
-        "holder_did_method": "key",
-    }
-
-    credential_response = await credo_client.post(
-        "/oid4vci/accept-offer", json=accept_offer_request
-    )
-    credo_credential = extract_credential(credential_response, "Credo")
-
-    # Step 2: Create verification request (using patterns compatible with both wallets)
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA", "ES256"]}},
-        "input_descriptors": [
-            {
-                "id": "cross-wallet-descriptor",
-                "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA", "ES256"]}},
-                "constraints": {
-                    "fields": [
-                        {
-                            "path": ["$.vct", "$.type"],
-                            "filter": {
-                                "type": "string",
-                                "const": "CrossWalletCredential",
-                            },
-                        },
-                        {"path": ["$.name", "$.credentialSubject.name"]},
-                    ]
-                },
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-    pres_def_id = pres_def_response["pres_def_id"]
-
-    presentation_request = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA", "ES256"]}},
-        },
-    )
-    request_uri = presentation_request["request_uri"]
-    presentation_id = presentation_request["presentation"]["presentation_id"]
-
-    # Step 3: Credo presents the credential
-    present_request = {"request_uri": request_uri, "credentials": [credo_credential]}
-    presentation_response = await credo_client.post(
-        "/oid4vp/present", json=present_request
-    )
-
-    assert (
-        presentation_response.status_code == 200
-    ), f"Presentation failed: {presentation_response.text}"
-    presentation_result = presentation_response.json()
-    assert presentation_result.get("success") is True
-
-    # Step 4: Verify ACA-Py received and validated
-    for _ in range(10):
-        latest = await acapy_verifier_admin.get(
-            f"/oid4vp/presentation/{presentation_id}"
-        )
-        if latest.get("state") == "presentation-valid":
-            break
-        await asyncio.sleep(1)
-    else:
-        pytest.fail(f"Presentation not validated. Final state: {latest.get('state')}")
-
-
-@pytest.mark.asyncio
-async def test_issue_to_sphereon_verify_with_credo_jwt_vc(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    credo_client,  # noqa: ARG001
-    sphereon_client,
-):
-    """Issue JWT VC to Sphereon, then try to verify if Credo can handle similar patterns.
-
-    This tests format compatibility between wallets for JWT VC credentials.
-    """
-    # Step 1: Issue JWT VC to Sphereon
-    random_suffix = str(uuid.uuid4())[:8]
-    cred_id = f"SphereonIssuedCredential-{random_suffix}"
-
-    supported = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create/jwt",
-        json={
-            "cryptographic_binding_methods_supported": ["did"],
-            "cryptographic_suites_supported": ["ES256"],
-            "format": "jwt_vc_json",
-            "id": cred_id,
-            "@context": [
-                "https://www.w3.org/2018/credentials/v1",
-                "https://www.w3.org/2018/credentials/examples/v1",
-            ],
-            "type": ["VerifiableCredential", "UniversityDegreeCredential"],
-        },
-    )
-    supported_cred_id = supported["supported_cred_id"]
-
-    did_result = await acapy_issuer_admin.post(
-        "/did/jwk/create", json={"key_type": "p256"}
-    )
-    issuer_did = did_result["did"]
-
-    exchange = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": supported_cred_id,
-            "credential_subject": {"name": "sphereon_test_user"},
-            "verification_method": issuer_did + "#0",
-        },
-    )
-
-    offer_response = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer",
-        params={"exchange_id": exchange["exchange_id"]},
-    )
-    credential_offer = offer_response["credential_offer"]
-
-    # Sphereon accepts offer
-    response = await sphereon_client.post(
-        "/oid4vci/accept-offer", json={"offer": credential_offer}
-    )
-    sphereon_credential = extract_credential(response, "Sphereon")
-
-    # Step 2: Create presentation definition for JWT VP
-    # NOTE: Using schema-based definition (like existing Sphereon tests)
-    # instead of format+constraints pattern which may cause interop issues
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "input_descriptors": [
-            {
-                "id": "university_degree",
-                "name": "University Degree",
-                "schema": [{"uri": "https://www.w3.org/2018/credentials/examples/v1"}],
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-    pres_def_id = pres_def_response["pres_def_id"]
-
-    request_response = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"jwt_vp_json": {"alg": ["ES256"]}},
-        },
-    )
-    request_uri = request_response["request_uri"]
-    presentation_id = request_response["presentation"]["presentation_id"]
-
-    # Step 3: Sphereon presents the credential
-    present_response = await sphereon_client.post(
-        "/oid4vp/present-credential",
-        json={
-            "authorization_request_uri": request_uri,
-            "verifiable_credentials": [sphereon_credential],
-        },
-    )
-    assert (
-        present_response.status_code == 200
-    ), f"Sphereon present failed: {present_response.text}"
-
-    # Step 4: Verify on ACA-Py side
-    record = None
-    for _ in range(10):
-        record = await acapy_verifier_admin.get(
-            f"/oid4vp/presentation/{presentation_id}"
-        )
-        if record["state"] == "presentation-valid":
-            break
-        await asyncio.sleep(1)
-    else:
-        # Capture diagnostic info for debugging the interop bug
-        error_info = {
-            "state": record.get("state") if record else "no record",
-            "errors": record.get("errors") if record else None,
-            "verified": record.get("verified") if record else None,
-        }
-        pytest.fail(
-            f"Sphereon JWT VP presentation rejected by ACA-Py verifier.\n"
-            f"This is an interoperability bug between Sphereon and ACA-Py OID4VP.\n"
-            f"Diagnostic info: {error_info}\n"
-            f"Credential format: jwt_vc_json, VP format: jwt_vp_json"
-        )
-
-
-@pytest.mark.asyncio
-@pytest.mark.xfail(
-    reason="Known bug: Sphereon VP with format+constraints pattern rejected by ACA-Py"
-)
-async def test_sphereon_jwt_vp_with_constraints_pattern(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    sphereon_client,
-):
-    """Test Sphereon JWT VP with format+constraints presentation definition.
-
-    KNOWN BUG: When using 'format' and 'constraints' in input_descriptors
-    instead of 'schema', Sphereon's VP is rejected by ACA-Py verifier.
-
-    This test documents the interoperability issue for future fixes.
-    """
-    random_suffix = str(uuid.uuid4())[:8]
-    cred_id = f"ConstraintsBugTest-{random_suffix}"
-
-    # Issue JWT VC to Sphereon
-    supported = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create/jwt",
-        json={
-            "cryptographic_binding_methods_supported": ["did"],
-            "cryptographic_suites_supported": ["ES256"],
-            "format": "jwt_vc_json",
-            "id": cred_id,
-            "@context": [
-                "https://www.w3.org/2018/credentials/v1",
-                "https://www.w3.org/2018/credentials/examples/v1",
-            ],
-            "type": ["VerifiableCredential", "TestCredential"],
-        },
-    )
-
-    did_result = await acapy_issuer_admin.post(
-        "/did/jwk/create", json={"key_type": "p256"}
-    )
-    issuer_did = did_result["did"]
-
-    exchange = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": supported["supported_cred_id"],
-            "credential_subject": {"test": "value"},
-            "verification_method": issuer_did + "#0",
-        },
-    )
-
-    offer_response = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange["exchange_id"]}
-    )
-
-    response = await sphereon_client.post(
-        "/oid4vci/accept-offer", json={"offer": offer_response["credential_offer"]}
-    )
-    credential = extract_credential(response, "Sphereon")
-
-    # Use format+constraints pattern (known to fail)
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "input_descriptors": [
-            {
-                "id": "test-descriptor",
-                "name": "Test Credential",
-                "format": {"jwt_vp_json": {"alg": ["ES256"]}},
-                "constraints": {
-                    "fields": [
-                        {
-                            "path": ["$.type"],
-                            "filter": {
-                                "type": "array",
-                                "contains": {"const": "TestCredential"},
-                            },
-                        },
-                    ]
-                },
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-
-    request_response = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_response["pres_def_id"],
-            "vp_formats": {"jwt_vp_json": {"alg": ["ES256"]}},
-        },
-    )
-
-    present_response = await sphereon_client.post(
-        "/oid4vp/present-credential",
-        json={
-            "authorization_request_uri": request_response["request_uri"],
-            "verifiable_credentials": [credential],
-        },
-    )
-    assert present_response.status_code == 200
-
-    # This should fail - documenting the bug
-    presentation_id = request_response["presentation"]["presentation_id"]
-    for _ in range(10):
-        record = await acapy_verifier_admin.get(
-            f"/oid4vp/presentation/{presentation_id}"
-        )
-        if record["state"] == "presentation-valid":
-            break
-        await asyncio.sleep(1)
-    else:
-        pytest.fail(
-            f"Expected failure: format+constraints pattern rejected. State: {record['state']}"
-        )
-
-
-# =============================================================================
-# Format Negotiation Edge Cases
-# =============================================================================
-
-
-@pytest.mark.asyncio
-async def test_credo_unsupported_algorithm_request(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    credo_client,
-):
-    """Test Credo behavior when verifier requests unsupported algorithm.
-
-    Issue credential with EdDSA, but request presentation with only ES256.
-    This tests algorithm negotiation handling.
-    """
-    random_suffix = str(uuid.uuid4())[:8]
-    credential_supported = {
-        "id": f"AlgoTestCredential_{random_suffix}",
-        "format": "vc+sd-jwt",
-        "scope": "AlgoTest",
-        "proof_types_supported": {
-            "jwt": {"proof_signing_alg_values_supported": ["EdDSA"]}  # EdDSA only
-        },
-        "format_data": {
-            "cryptographic_binding_methods_supported": ["did:key"],
-            "cryptographic_suites_supported": ["EdDSA"],
-            "vct": "AlgoTestCredential",
-            "claims": {"test_field": {"mandatory": True}},
-        },
-        "vc_additional_data": {"sd_list": ["/test_field"]},
-    }
-
-    config_response = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create", json=credential_supported
-    )
-    config_id = config_response["supported_cred_id"]
-
-    did_response = await acapy_issuer_admin.post(
-        "/wallet/did/create", json={"method": "key", "options": {"key_type": "ed25519"}}
-    )
-    issuer_did = did_response["result"]["did"]
-
-    exchange_response = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": config_id,
-            "credential_subject": {"test_field": "algo_test_value"},
-            "did": issuer_did,
-        },
-    )
-    exchange_id = exchange_response["exchange_id"]
-
-    offer_response = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange_id}
-    )
-
-    # Credo accepts offer
-    credo_response = await credo_client.post(
-        "/oid4vci/accept-offer",
-        json={
-            "credential_offer": offer_response["credential_offer"],
-            "holder_did_method": "key",
-        },
-    )
-    credo_credential = extract_credential(credo_response, "Credo")
-
-    # Create verification request that ONLY accepts ES256 (not EdDSA)
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["ES256"]}},  # ES256 only
-        "input_descriptors": [
-            {
-                "id": "algo-test",
-                "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["ES256"]}},
-                "constraints": {
-                    "fields": [
-                        {
-                            "path": ["$.vct"],
-                            "filter": {"type": "string", "const": "AlgoTestCredential"},
-                        },
-                    ]
-                },
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-    pres_def_id = pres_def_response["pres_def_id"]
-
-    presentation_request = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"vc+sd-jwt": {"sd-jwt_alg_values": ["ES256"]}},
-        },
-    )
-    request_uri = presentation_request["request_uri"]
-
-    # Attempt presentation - this should either fail or Credo should handle algorithm mismatch
-    present_response = await credo_client.post(
-        "/oid4vp/present",
-        json={"request_uri": request_uri, "credentials": [credo_credential]},
-    )
-
-    # Document the behavior - this test discovers if there's a bug
-    # Expected: Either Credo rejects with meaningful error, or verifier rejects the presentation
-    if present_response.status_code == 200:
-        # If presentation was attempted, check verifier's response
-        result = present_response.json()
-        # The presentation may have been submitted but should fail verification
-        if result.get("success") is True:
-            # Check if ACA-Py correctly rejects the mismatched algorithm
-            presentation_id = presentation_request["presentation"]["presentation_id"]
-            for _ in range(5):
-                record = await acapy_verifier_admin.get(
-                    f"/oid4vp/presentation/{presentation_id}"
-                )
-                if record.get("state") in [
-                    "presentation-valid",
-                    "presentation-invalid",
-                ]:
-                    break
-                await asyncio.sleep(1)
-
-            # Document the actual behavior for bug discovery
-            print(f"Algorithm mismatch test result: state={record.get('state')}")
-            # If state is "presentation-valid", this indicates a potential bug where
-            # algorithm constraints are not being enforced
-    else:
-        # Credo correctly rejected the request
-        print(f"Credo rejected algorithm mismatch: {present_response.status_code}")
-
-
-@pytest.mark.asyncio
-async def test_sphereon_unsupported_format_request(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    sphereon_client,
-):
-    """Test Sphereon behavior when asked to present unsupported format.
-
-    Issue JWT VC but request SD-JWT presentation format.
-    """
-    random_suffix = str(uuid.uuid4())[:8]
-    cred_id = f"FormatTestCredential-{random_suffix}"
-
-    # Issue JWT VC (not SD-JWT)
-    supported = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create/jwt",
-        json={
-            "cryptographic_binding_methods_supported": ["did"],
-            "cryptographic_suites_supported": ["ES256"],
-            "format": "jwt_vc_json",
-            "id": cred_id,
-            "@context": ["https://www.w3.org/2018/credentials/v1"],
-            "type": ["VerifiableCredential", "TestCredential"],
-        },
-    )
-    supported_cred_id = supported["supported_cred_id"]
-
-    did_result = await acapy_issuer_admin.post(
-        "/did/jwk/create", json={"key_type": "p256"}
-    )
-    issuer_did = did_result["did"]
-
-    exchange = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": supported_cred_id,
-            "credential_subject": {"test": "value"},
-            "verification_method": issuer_did + "#0",
-        },
-    )
-
-    offer_response = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange["exchange_id"]}
-    )
-
-    # Sphereon accepts JWT VC
-    response = await sphereon_client.post(
-        "/oid4vci/accept-offer", json={"offer": offer_response["credential_offer"]}
-    )
-    jwt_credential = extract_credential(response, "Sphereon")
-
-    # Create request for SD-JWT format (mismatched)
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["ES256"]}},  # SD-JWT, not JWT VC
-        "input_descriptors": [
-            {
-                "id": "format-test",
-                "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["ES256"]}},
-                "constraints": {"fields": [{"path": ["$.vct"]}]},
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-    pres_def_id = pres_def_response["pres_def_id"]
-
-    request_response = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"vc+sd-jwt": {"sd-jwt_alg_values": ["ES256"]}},
-        },
-    )
-    request_uri = request_response["request_uri"]
-
-    # Attempt to present JWT VC as SD-JWT - should fail
-    present_response = await sphereon_client.post(
-        "/oid4vp/present-credential",
-        json={
-            "authorization_request_uri": request_uri,
-            "verifiable_credentials": [jwt_credential],
-        },
-    )
-
-    # Document behavior for bug discovery
-    print(f"Format mismatch test: Sphereon returned {present_response.status_code}")
-    if present_response.status_code == 200:
-        print("WARNING: Sphereon accepted format mismatch - potential interop issue")
-    else:
-        print(f"Sphereon correctly rejected: {present_response.text}")
-
-
-# =============================================================================
-# Selective Disclosure Parity Tests
-# =============================================================================
-
-
-@pytest.mark.asyncio
-async def test_selective_disclosure_credo_vs_sphereon_parity(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    credo_client,
-):
-    """Test selective disclosure behavior in Credo matches expected behavior.
-
-    Issue SD-JWT with multiple disclosable claims, request only subset,
-    verify only requested claims are disclosed.
-    """
-    random_suffix = str(uuid.uuid4())[:8]
-    credential_supported = {
-        "id": f"SDTestCredential_{random_suffix}",
-        "format": "vc+sd-jwt",
-        "scope": "SDTest",
-        "proof_types_supported": {
-            "jwt": {"proof_signing_alg_values_supported": ["EdDSA", "ES256"]}
-        },
-        "format_data": {
-            "cryptographic_binding_methods_supported": ["did:key"],
-            "cryptographic_suites_supported": ["EdDSA"],
-            "vct": "SDTestCredential",
-            "claims": {
-                "public_claim": {"mandatory": True},
-                "private_claim_1": {"mandatory": False},
-                "private_claim_2": {"mandatory": False},
-                "private_claim_3": {"mandatory": False},
-            },
-        },
-        "vc_additional_data": {
-            "sd_list": ["/private_claim_1", "/private_claim_2", "/private_claim_3"]
-        },
-    }
-
-    config_response = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create", json=credential_supported
-    )
-    config_id = config_response["supported_cred_id"]
-
-    did_response = await acapy_issuer_admin.post(
-        "/wallet/did/create", json={"method": "key", "options": {"key_type": "ed25519"}}
-    )
-    issuer_did = did_response["result"]["did"]
-
-    exchange_response = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": config_id,
-            "credential_subject": {
-                "public_claim": "public_value",
-                "private_claim_1": "secret_1",
-                "private_claim_2": "secret_2",
-                "private_claim_3": "secret_3",
-            },
-            "did": issuer_did,
-        },
-    )
-    exchange_id = exchange_response["exchange_id"]
-
-    offer_response = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange_id}
-    )
-
-    # Credo accepts
-    credo_response = await credo_client.post(
-        "/oid4vci/accept-offer",
-        json={
-            "credential_offer": offer_response["credential_offer"],
-            "holder_did_method": "key",
-        },
-    )
-    sd_jwt_credential = extract_credential(credo_response, "Credo")
-
-    # Request ONLY private_claim_1 (not 2 or 3)
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA", "ES256"]}},
-        "input_descriptors": [
-            {
-                "id": "sd-test",
-                "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA", "ES256"]}},
-                "constraints": {
-                    "limit_disclosure": "required",
-                    "fields": [
-                        {
-                            "path": ["$.vct"],
-                            "filter": {"type": "string", "const": "SDTestCredential"},
-                        },
-                        {
-                            "path": [
-                                "$.private_claim_1",
-                                "$.credentialSubject.private_claim_1",
-                            ]
-                        },
-                        # NOT requesting private_claim_2 or private_claim_3
-                    ],
-                },
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-    pres_def_id = pres_def_response["pres_def_id"]
-
-    presentation_request = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA", "ES256"]}},
-        },
-    )
-    request_uri = presentation_request["request_uri"]
-    presentation_id = presentation_request["presentation"]["presentation_id"]
-
-    # Credo presents with selective disclosure
-    present_response = await credo_client.post(
-        "/oid4vp/present",
-        json={"request_uri": request_uri, "credentials": [sd_jwt_credential]},
-    )
-    assert (
-        present_response.status_code == 200
-    ), f"Present failed: {present_response.text}"
-
-    # Verify presentation and check disclosed claims
-    for _ in range(10):
-        record = await acapy_verifier_admin.get(
-            f"/oid4vp/presentation/{presentation_id}"
-        )
-        if record.get("state") in ["presentation-valid", "presentation-invalid"]:
-            break
-        await asyncio.sleep(1)
-
-    assert record.get("state") == "presentation-valid", f"Failed: {record.get('state')}"
-
-    # Check what was disclosed in the verified claims
-    verified_claims = record.get("verified_claims", {})
-    print(f"Selective disclosure test - verified claims: {verified_claims}")
-
-    # Bug discovery: Check if unrequested claims were incorrectly disclosed
-    if verified_claims:
-        # These should NOT be present if selective disclosure is working correctly
-        if "private_claim_2" in str(verified_claims) or "private_claim_3" in str(
-            verified_claims
-        ):
-            print("WARNING: Unrequested claims were disclosed - potential SD bug")
-
-
-@pytest.mark.asyncio
-async def test_selective_disclosure_all_claims_disclosed(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    credo_client,
-):
-    """Test that all requested claims ARE disclosed when requested."""
-    random_suffix = str(uuid.uuid4())[:8]
-    credential_supported = {
-        "id": f"FullSDCredential_{random_suffix}",
-        "format": "vc+sd-jwt",
-        "scope": "FullSDTest",
-        "proof_types_supported": {
-            "jwt": {"proof_signing_alg_values_supported": ["EdDSA"]}
-        },
-        "format_data": {
-            "cryptographic_binding_methods_supported": ["did:key"],
-            "cryptographic_suites_supported": ["EdDSA"],
-            "vct": "FullSDCredential",
-            "claims": {
-                "claim_a": {"mandatory": True},
-                "claim_b": {"mandatory": True},
-                "claim_c": {"mandatory": True},
-            },
-        },
-        "vc_additional_data": {"sd_list": ["/claim_a", "/claim_b", "/claim_c"]},
-    }
-
-    config_response = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create", json=credential_supported
-    )
-    config_id = config_response["supported_cred_id"]
-
-    did_response = await acapy_issuer_admin.post(
-        "/wallet/did/create", json={"method": "key", "options": {"key_type": "ed25519"}}
-    )
-    issuer_did = did_response["result"]["did"]
-
-    exchange_response = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": config_id,
-            "credential_subject": {
-                "claim_a": "value_a",
-                "claim_b": "value_b",
-                "claim_c": "value_c",
-            },
-            "did": issuer_did,
-        },
-    )
-
-    offer_response = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer",
-        params={"exchange_id": exchange_response["exchange_id"]},
-    )
-
-    credo_response = await credo_client.post(
-        "/oid4vci/accept-offer",
-        json={
-            "credential_offer": offer_response["credential_offer"],
-            "holder_did_method": "key",
-        },
-    )
-    credential = extract_credential(credo_response, "Credo")
-
-    # Request ALL claims
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA"]}},
-        "input_descriptors": [
-            {
-                "id": "full-sd-test",
-                "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA"]}},
-                "constraints": {
-                    "limit_disclosure": "required",
-                    "fields": [
-                        {"path": ["$.vct"], "filter": {"const": "FullSDCredential"}},
-                        {"path": ["$.claim_a", "$.credentialSubject.claim_a"]},
-                        {"path": ["$.claim_b", "$.credentialSubject.claim_b"]},
-                        {"path": ["$.claim_c", "$.credentialSubject.claim_c"]},
-                    ],
-                },
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-    pres_def_id = pres_def_response["pres_def_id"]
-
-    presentation_request = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA"]}},
-        },
-    )
-    presentation_id = presentation_request["presentation"]["presentation_id"]
-
-    present_response = await credo_client.post(
-        "/oid4vp/present",
-        json={
-            "request_uri": presentation_request["request_uri"],
-            "credentials": [credential],
-        },
-    )
-    assert present_response.status_code == 200
-
-    for _ in range(10):
-        record = await acapy_verifier_admin.get(
-            f"/oid4vp/presentation/{presentation_id}"
-        )
-        if record.get("state") == "presentation-valid":
-            break
-        await asyncio.sleep(1)
-
-    assert record.get("state") == "presentation-valid"
-
-    # Verify all requested claims are present
-    verified_claims = record.get("verified_claims", {})
-    print(f"Full disclosure test - verified claims: {verified_claims}")
-
-
-# =============================================================================
-# mDOC Cross-Wallet Tests
-# =============================================================================
-
-
-@pytest.mark.skipif(not MDOC_AVAILABLE, reason="isomdl_uniffi not available")
-@pytest.mark.asyncio
-async def test_mdoc_issue_to_credo_verify_with_sphereon_patterns(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    credo_client,
-    sphereon_client,  # noqa: ARG001
-    setup_all_trust_anchors,  # noqa: ARG001 - Required for mDOC verification
-):
-    """Issue mDOC to Credo and verify using Sphereon-compatible verification patterns.
-
-    Tests mDOC format interoperability between wallets.
-    """
-    random_suffix = str(uuid.uuid4())[:8]
-    credential_supported = {
-        "id": f"MdocCrossWallet_{random_suffix}",
-        "format": "mso_mdoc",
-        "scope": "MdocCrossWalletTest",
-        "cryptographic_binding_methods_supported": ["cose_key", "did:key", "did"],
-        "cryptographic_suites_supported": ["ES256"],
-        "proof_types_supported": {
-            "jwt": {"proof_signing_alg_values_supported": ["ES256"]}
-        },
-        "format_data": {
-            "doctype": "org.iso.18013.5.1.mDL",
-            "claims": {
-                "org.iso.18013.5.1": {
-                    "given_name": {"mandatory": True},
-                    "family_name": {"mandatory": True},
-                }
-            },
-        },
-    }
-
-    config_response = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create", json=credential_supported
-    )
-    config_id = config_response["supported_cred_id"]
-
-    did_response = await acapy_issuer_admin.post(
-        "/wallet/did/create", json={"method": "key", "options": {"key_type": "p256"}}
-    )
-    issuer_did = did_response["result"]["did"]
-
-    exchange_response = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": config_id,
-            "credential_subject": {
-                "org.iso.18013.5.1": {
-                    "given_name": "Cross",
-                    "family_name": "Wallet",
-                }
-            },
-            "did": issuer_did,
-        },
-    )
-
-    offer_response = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer",
-        params={"exchange_id": exchange_response["exchange_id"]},
-    )
-
-    # Credo accepts mDOC
-    credo_response = await credo_client.post(
-        "/oid4vci/accept-offer",
-        json={
-            "credential_offer": offer_response["credential_offer"],
-            "holder_did_method": "key",
-        },
-    )
-    mdoc_credential = extract_credential(credo_response, "Credo")
-
-    # Verify format if response successful
-    result = credo_response.json()
-    if "format" in result:
-        assert result["format"] == "mso_mdoc"
-
-    # Create mDOC presentation request
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "format": {"mso_mdoc": {"alg": ["ES256"]}},
-        "input_descriptors": [
-            {
-                "id": "org.iso.18013.5.1.mDL",
-                "format": {"mso_mdoc": {"alg": ["ES256"]}},
-                "constraints": {
-                    "limit_disclosure": "required",
-                    "fields": [
-                        {
-                            "path": ["$['org.iso.18013.5.1']['given_name']"],
-                            "intent_to_retain": False,
-                        },
-                        {
-                            "path": ["$['org.iso.18013.5.1']['family_name']"],
-                            "intent_to_retain": False,
-                        },
-                    ],
-                },
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-    pres_def_id = pres_def_response["pres_def_id"]
-
-    presentation_request = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"mso_mdoc": {"alg": ["ES256"]}},
-        },
-    )
-    presentation_id = presentation_request["presentation"]["presentation_id"]
-
-    # Credo presents mDOC
-    present_response = await credo_client.post(
-        "/oid4vp/present",
-        json={
-            "request_uri": presentation_request["request_uri"],
-            "credentials": [mdoc_credential],
-        },
-    )
-    assert (
-        present_response.status_code == 200
-    ), f"Credo mDOC present failed: {present_response.text}"
-
-    # Verify on ACA-Py
-    for _ in range(10):
-        record = await acapy_verifier_admin.get(
-            f"/oid4vp/presentation/{presentation_id}"
-        )
-        if record.get("state") == "presentation-valid":
-            break
-        await asyncio.sleep(1)
-
-    assert (
-        record.get("state") == "presentation-valid"
-    ), f"mDOC verification failed: {record.get('state')}"
-    print("mDOC cross-wallet test passed!")
-
-
-@pytest.mark.skipif(not MDOC_AVAILABLE, reason="isomdl_uniffi not available")
-@pytest.mark.asyncio
-async def test_mdoc_issue_to_sphereon_verify_with_credo_patterns(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    sphereon_client,
-    setup_all_trust_anchors,  # noqa: ARG001 - Required for mDOC verification
-):
-    """Issue mDOC to Sphereon and verify.
-
-    Tests Sphereon's mDOC handling and verification compatibility.
-    """
-    random_suffix = str(uuid.uuid4())[:8]
-    cred_id = f"mDL-Sphereon-{random_suffix}"
-
-    supported = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create",
-        json={
-            "cryptographic_binding_methods_supported": ["cose_key"],
-            "cryptographic_suites_supported": ["ES256"],
-            "format": "mso_mdoc",
-            "id": cred_id,
-            "identifier": "org.iso.18013.5.1.mDL",
-            "format_data": {"doctype": "org.iso.18013.5.1.mDL"},
-            "claims": {
-                "org.iso.18013.5.1": {
-                    "given_name": {"mandatory": True},
-                    "family_name": {"mandatory": True},
-                }
-            },
-        },
-    )
-    supported_cred_id = supported["supported_cred_id"]
-
-    did_result = await acapy_issuer_admin.post(
-        "/did/jwk/create", json={"key_type": "p256"}
-    )
-    issuer_did = did_result["did"]
-
-    exchange = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": supported_cred_id,
-            "credential_subject": {
-                "org.iso.18013.5.1": {
-                    "given_name": "Sphereon",
-                    "family_name": "Test",
-                }
-            },
-            "verification_method": issuer_did + "#0",
-        },
-    )
-
-    offer_response = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange["exchange_id"]}
-    )
-
-    # Sphereon accepts mDOC
-    response = await sphereon_client.post(
-        "/oid4vci/accept-offer",
-        json={"offer": offer_response["credential_offer"], "format": "mso_mdoc"},
-    )
-    mdoc_credential = extract_credential(response, "Sphereon")
-
-    # Create mDOC presentation request
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "input_descriptors": [
-            {
-                "id": "mdl",
-                "format": {"mso_mdoc": {"alg": ["ES256"]}},
-                "constraints": {
-                    "limit_disclosure": "required",
-                    "fields": [
-                        {
-                            "path": ["$['org.iso.18013.5.1']['given_name']"],
-                            "intent_to_retain": False,
-                        },
-                    ],
-                },
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-    pres_def_id = pres_def_response["pres_def_id"]
-
-    request_response = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"mso_mdoc": {"alg": ["ES256"]}},
-        },
-    )
-    presentation_id = request_response["presentation"]["presentation_id"]
-
-    # Sphereon presents
-    present_response = await sphereon_client.post(
-        "/oid4vp/present-credential",
-        json={
-            "authorization_request_uri": request_response["request_uri"],
-            "verifiable_credentials": [mdoc_credential],
-        },
-    )
-    assert (
-        present_response.status_code == 200
-    ), f"Sphereon mDOC present failed: {present_response.text}"
-
-    # Verify
-    for _ in range(10):
-        record = await acapy_verifier_admin.get(
-            f"/oid4vp/presentation/{presentation_id}"
-        )
-        if record.get("state") == "presentation-valid":
-            break
-        await asyncio.sleep(1)
-
-    assert (
-        record.get("state") == "presentation-valid"
-    ), f"Sphereon mDOC verification failed: {record.get('state')}"
-
-
-# =============================================================================
-# Multi-Credential Presentation Tests
-# =============================================================================
-
-
-@pytest.mark.asyncio
-async def test_credo_multi_credential_presentation(
-    acapy_issuer_admin,
-    acapy_verifier_admin,
-    credo_client,
-):
-    """Test Credo presenting multiple credentials in a single presentation.
-
-    This tests whether multi-credential flows work correctly.
-    """
-    random_suffix = str(uuid.uuid4())[:8]
-
-    # Create two different credential types
-    cred_config_1 = {
-        "id": f"IdentityCredential_{random_suffix}",
-        "format": "vc+sd-jwt",
-        "scope": "Identity",
-        "proof_types_supported": {
-            "jwt": {"proof_signing_alg_values_supported": ["EdDSA"]}
-        },
-        "format_data": {
-            "cryptographic_binding_methods_supported": ["did:key"],
-            "cryptographic_suites_supported": ["EdDSA"],
-            "vct": "IdentityCredential",
-            "claims": {"name": {"mandatory": True}},
-        },
-        "vc_additional_data": {"sd_list": ["/name"]},
-    }
-
-    cred_config_2 = {
-        "id": f"EmploymentCredential_{random_suffix}",
-        "format": "vc+sd-jwt",
-        "scope": "Employment",
-        "proof_types_supported": {
-            "jwt": {"proof_signing_alg_values_supported": ["EdDSA"]}
-        },
-        "format_data": {
-            "cryptographic_binding_methods_supported": ["did:key"],
-            "cryptographic_suites_supported": ["EdDSA"],
-            "vct": "EmploymentCredential",
-            "claims": {"employer": {"mandatory": True}},
-        },
-        "vc_additional_data": {"sd_list": ["/employer"]},
-    }
-
-    config_1 = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create", json=cred_config_1
-    )
-    config_2 = await acapy_issuer_admin.post(
-        "/oid4vci/credential-supported/create", json=cred_config_2
-    )
-
-    did_response = await acapy_issuer_admin.post(
-        "/wallet/did/create", json={"method": "key", "options": {"key_type": "ed25519"}}
-    )
-    issuer_did = did_response["result"]["did"]
-
-    # Issue credential 1
-    exchange_1 = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": config_1["supported_cred_id"],
-            "credential_subject": {"name": "Multi Test User"},
-            "did": issuer_did,
-        },
-    )
-    offer_1 = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange_1["exchange_id"]}
-    )
-    credo_resp_1 = await credo_client.post(
-        "/oid4vci/accept-offer",
-        json={
-            "credential_offer": offer_1["credential_offer"],
-            "holder_did_method": "key",
-        },
-    )
-    credential_1 = extract_credential(credo_resp_1, "Credo")
-
-    # Issue credential 2
-    exchange_2 = await acapy_issuer_admin.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": config_2["supported_cred_id"],
-            "credential_subject": {"employer": "Test Corp"},
-            "did": issuer_did,
-        },
-    )
-    offer_2 = await acapy_issuer_admin.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange_2["exchange_id"]}
-    )
-    credo_resp_2 = await credo_client.post(
-        "/oid4vci/accept-offer",
-        json={
-            "credential_offer": offer_2["credential_offer"],
-            "holder_did_method": "key",
-        },
-    )
-    credential_2 = extract_credential(credo_resp_2, "Credo")
-
-    # Create presentation definition requesting BOTH credentials
-    presentation_definition = {
-        "id": str(uuid.uuid4()),
-        "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA"]}},
-        "input_descriptors": [
-            {
-                "id": "identity-descriptor",
-                "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA"]}},
-                "constraints": {
-                    "fields": [
-                        {"path": ["$.vct"], "filter": {"const": "IdentityCredential"}},
-                        {"path": ["$.name", "$.credentialSubject.name"]},
-                    ]
-                },
-            },
-            {
-                "id": "employment-descriptor",
-                "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA"]}},
-                "constraints": {
-                    "fields": [
-                        {
-                            "path": ["$.vct"],
-                            "filter": {"const": "EmploymentCredential"},
-                        },
-                        {"path": ["$.employer", "$.credentialSubject.employer"]},
-                    ]
-                },
-            },
-        ],
-    }
-
-    pres_def_response = await acapy_verifier_admin.post(
-        "/oid4vp/presentation-definition", json={"pres_def": presentation_definition}
-    )
-    pres_def_id = pres_def_response["pres_def_id"]
-
-    presentation_request = await acapy_verifier_admin.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA"]}},
-        },
-    )
-    presentation_id = presentation_request["presentation"]["presentation_id"]
-
-    # Credo presents BOTH credentials
-    present_response = await credo_client.post(
-        "/oid4vp/present",
-        json={
-            "request_uri": presentation_request["request_uri"],
-            "credentials": [credential_1, credential_2],
-        },
-    )
-
-    # Document behavior
-    print(f"Multi-credential presentation status: {present_response.status_code}")
-    if present_response.status_code == 200:
-        result = present_response.json()
-        print(f"Multi-credential result: {result}")
-
-        # Check verification
-        for _ in range(10):
-            record = await acapy_verifier_admin.get(
-                f"/oid4vp/presentation/{presentation_id}"
-            )
-            if record.get("state") in ["presentation-valid", "presentation-invalid"]:
-                break
-            await asyncio.sleep(1)
-
-        print(f"Multi-credential verification state: {record.get('state')}")
-        if record.get("state") != "presentation-valid":
-            print("WARNING: Multi-credential presentation failed - potential bug")
-    else:
-        print(f"Multi-credential presentation failed: {present_response.text}")
