--- oid4vc/integration/tests/test_interop/test_acapy_credo_flow.py
@@ -1,269 +1,0 @@
-"""Test ACA-Py ↔ Credo OID4VC flow.
-
-Tests the complete flow:
-1. ACA-Py issuer creates credential offer
-2. Credo accepts credential from ACA-Py issuer (OID4VCI)
-3. Credo presents credential to ACA-Py verifier (OID4VP)
-4. ACA-Py verifier validates presentation
-"""
-
-
-import httpx
-import pytest
-
-
-@pytest.mark.asyncio
-async def test_acapy_to_credo_to_acapy_flow(
-    acapy_issuer: httpx.AsyncClient, acapy_verifier: httpx.AsyncClient, credo
-):
-    """Test complete flow: ACA-Py issuer → Credo → ACA-Py verifier."""
-
-    # Step 1: Check that all services are healthy
-    issuer_status = await acapy_issuer.get("/status/ready")
-    assert issuer_status.status_code == 200, "ACA-Py issuer is not ready"
-
-    verifier_status = await acapy_verifier.get("/status/ready")
-    assert verifier_status.status_code == 200, "ACA-Py verifier is not ready"
-
-    # Test basic Credo connectivity
-    credo_test = await credo.test()
-    assert credo_test is not None, "Credo is not responding"
-
-    print("✅ All services are healthy")
-
-
-@pytest.mark.asyncio
-async def test_credential_issuance_flow(acapy_issuer: httpx.AsyncClient, credo):
-    """Test credential issuance from ACA-Py to Credo."""
-
-    # Step 1: Create a supported credential type on ACA-Py issuer
-    import uuid
-
-    unique_id = f"TestCredential_{uuid.uuid4().hex[:8]}"
-
-    supported_cred_request = {
-        "id": unique_id,
-        "format": "jwt_vc_json",
-        "format_data": {
-            "types": ["VerifiableCredential", "TestCredential"],
-            "credentialSubject": {
-                "name": {"display": [{"name": "Full Name", "locale": "en-US"}]},
-                "email": {"display": [{"name": "Email Address", "locale": "en-US"}]},
-            },
-        },
-        "cryptographic_binding_methods_supported": ["did"],
-        "cryptographic_suites_supported": ["ES256K"],
-        "display": [
-            {
-                "name": "Test Credential",
-                "locale": "en-US",
-                "background_color": "#12107c",
-                "text_color": "#FFFFFF",
-            }
-        ],
-    }
-
-    print("📝 Creating supported credential...")
-    response = await acapy_issuer.post(
-        "/oid4vci/credential-supported/create", json=supported_cred_request
-    )
-    print(f"Supported credential response: {response.status_code}")
-    if response.status_code != 200:
-        print(f"Response body: {response.text}")
-    assert (
-        response.status_code == 200
-    ), f"Failed to create supported credential: {response.text}"
-
-    supported_cred = response.json()
-    supported_cred_id = supported_cred["supported_cred_id"]
-    print(f"✅ Created supported credential with ID: {supported_cred_id}")
-
-    # Step 2: Create credential exchange record
-    # Create a DID for the issuer
-    did_response = await acapy_issuer.post(
-        "/wallet/did/create", json={"method": "key", "options": {"key_type": "ed25519"}}
-    )
-    assert did_response.status_code == 200, f"Failed to create DID: {did_response.text}"
-    issuer_did = did_response.json()["result"]["did"]
-
-    exchange_request = {
-        "supported_cred_id": supported_cred_id,
-        "credential_subject": {"name": "John Doe", "email": "john.doe@example.com"},
-        "did": issuer_did,
-    }
-
-    print("🔄 Creating credential exchange...")
-    response = await acapy_issuer.post(
-        "/oid4vci/exchange/create", json=exchange_request
-    )
-    print(f"Exchange creation response: {response.status_code}")
-    if response.status_code != 200:
-        print(f"Response body: {response.text}")
-    assert response.status_code == 200, f"Failed to create exchange: {response.text}"
-
-    exchange = response.json()
-    exchange_id = exchange["exchange_id"]
-    print(f"✅ Created exchange with ID: {exchange_id}")
-
-    # Step 3: Get credential offer
-    print("📋 Getting credential offer...")
-    response = await acapy_issuer.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange_id}
-    )
-    print(f"Credential offer response: {response.status_code}")
-    if response.status_code != 200:
-        print(f"Response body: {response.text}")
-    assert (
-        response.status_code == 200
-    ), f"Failed to get credential offer: {response.text}"
-
-    offer_response = response.json()
-    print(f"✅ Got credential offer: {offer_response.keys()}")
-    print(f"📋 Credential offer content: {offer_response.get('credential_offer')}")
-    print(f"📋 Credential offer URI: {offer_response.get('credential_offer_uri')}")
-
-    # Step 4: Have Credo accept the credential offer
-    print("🤝 Having Credo accept the credential offer...")
-    try:
-        credo_result = await credo.openid4vci_accept_offer(
-            offer_response.get("credential_offer")
-        )
-        print(f"✅ Credo accepted credential offer: {credo_result}")
-    except Exception as e:
-        print(f"❌ Credo failed to accept offer: {e}")
-        # For now, let's not fail the test - just log the issue
-        print("📝 Note: Credo integration needs further work")
-
-    print("✅ Credential issuance flow completed")
-
-
-@pytest.mark.asyncio
-async def test_presentation_verification_flow(
-    acapy_issuer: httpx.AsyncClient,
-    acapy_verifier: httpx.AsyncClient,
-    credo,
-):
-    """Test presentation from Credo to ACA-Py verifier.
-
-    Complete flow:
-    1. Issue SD-JWT credential from ACA-Py to Credo
-    2. Create presentation request on ACA-Py verifier
-    3. Credo presents credential to ACA-Py verifier
-    4. Verify presentation is valid
-    """
-    import asyncio
-    import uuid
-
-    # Step 1: Issue a credential to Credo first
-    random_suffix = str(uuid.uuid4())[:8]
-    credential_supported = {
-        "id": f"IdentityCredential_{random_suffix}",
-        "format": "vc+sd-jwt",
-        "scope": "IdentityCredential",
-        "proof_types_supported": {
-            "jwt": {"proof_signing_alg_values_supported": ["EdDSA", "ES256"]}
-        },
-        "format_data": {
-            "cryptographic_binding_methods_supported": ["did:key"],
-            "cryptographic_suites_supported": ["EdDSA"],
-            "vct": "IdentityCredential",
-            "claims": {
-                "given_name": {"mandatory": True},
-                "family_name": {"mandatory": True},
-            },
-        },
-        "vc_additional_data": {"sd_list": ["/given_name", "/family_name"]},
-    }
-
-    response = await acapy_issuer.post(
-        "/oid4vci/credential-supported/create", json=credential_supported
-    )
-    response.raise_for_status()
-    config_id = response.json()["supported_cred_id"]
-
-    did_response = await acapy_issuer.post(
-        "/wallet/did/create", json={"method": "key", "options": {"key_type": "ed25519"}}
-    )
-    did_response.raise_for_status()
-    issuer_did = did_response.json()["result"]["did"]
-
-    exchange_response = await acapy_issuer.post(
-        "/oid4vci/exchange/create",
-        json={
-            "supported_cred_id": config_id,
-            "credential_subject": {"given_name": "Alice", "family_name": "Smith"},
-            "did": issuer_did,
-        },
-    )
-    exchange_response.raise_for_status()
-    exchange_id = exchange_response.json()["exchange_id"]
-
-    offer_response = await acapy_issuer.get(
-        "/oid4vci/credential-offer", params={"exchange_id": exchange_id}
-    )
-    offer_response.raise_for_status()
-    credential_offer = offer_response.json()["credential_offer"]
-
-    # Have Credo accept the credential
-    credo_credential = await credo.openid4vci_accept_offer(credential_offer)
-    print(f"✅ Credo received credential: {credo_credential.get('format', 'unknown')}")
-
-    # Step 2: Create presentation request on ACA-Py verifier
-    pres_def = {
-        "id": str(uuid.uuid4()),
-        "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA", "ES256"]}},
-        "input_descriptors": [
-            {
-                "id": "identity-descriptor",
-                "format": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA", "ES256"]}},
-                "constraints": {
-                    "fields": [
-                        {
-                            "path": ["$.vct"],
-                            "filter": {"type": "string", "const": "IdentityCredential"},
-                        }
-                    ]
-                },
-            }
-        ],
-    }
-
-    pres_def_response = await acapy_verifier.post(
-        "/oid4vp/presentation-definition", json={"pres_def": pres_def}
-    )
-    pres_def_response.raise_for_status()
-    pres_def_id = pres_def_response.json()["pres_def_id"]
-
-    request_response = await acapy_verifier.post(
-        "/oid4vp/request",
-        json={
-            "pres_def_id": pres_def_id,
-            "vp_formats": {"vc+sd-jwt": {"sd-jwt_alg_values": ["EdDSA", "ES256"]}},
-        },
-    )
-    request_response.raise_for_status()
-    request_data = request_response.json()
-    request_uri = request_data["request_uri"]
-    presentation_id = request_data["presentation"]["presentation_id"]
-    print(f"✅ Created presentation request: {request_uri}")
-
-    # Step 3: Have Credo present the credential
-    presentation_result = await credo.openid4vp_accept_request(request_uri)
-    print(f"✅ Credo submitted presentation: {presentation_result}")
-
-    # Step 4: Poll for presentation validation
-    for _ in range(15):
-        status_response = await acapy_verifier.get(
-            f"/oid4vp/presentation/{presentation_id}"
-        )
-        status_response.raise_for_status()
-        status = status_response.json()
-        if status.get("state") == "presentation-valid":
-            break
-        await asyncio.sleep(1.0)
-
-    assert (
-        status.get("state") == "presentation-valid"
-    ), f"Presentation not validated. Final state: {status.get('state')}"
-
-    print("✅ Presentation verification flow completed successfully!")
